+++ src/main/java/com/example/model/entities/SeriesLog.java
@@ -1,0 +1,43 @@
+package com.example.model.entities;
+
+import jakarta.persistence.*;
+
+import java.io.Serializable;
+
+@Entity
+@Table(name = "series_log")
+public class SeriesLog implements Serializable {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "log_description")
+    private String description;
+
+
+    public SeriesLog() {
+        // Required empty constructor
+    }
+
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public String toString() {
+        return "SeriesLog{" + "id=" + id + ", description='" + description + '\'' + '}';
+    }
+}
