+++ src/main/java/com/example/services/SeriesService.java
@@ -1,0 +1,6 @@
+package com.example.services;
+
+public interface SeriesService {
+
+    String getSeries();
+}
