--- src/main/java/model/entities/Actor.java
@@ -1,65 +1,0 @@
-package model.entities;
-
-
-import jakarta.persistence.*;
-
-import java.io.Serializable;
-import java.util.List;
-
-@Entity
-@Table(name = "actor")
-public class Actor implements Serializable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "name", unique = true)
-    private String name;
-
-    // Actor -> Series
-    @ManyToMany(mappedBy = "actorsList")
-    private List<Series> seriesList;
-
-
-    public Actor() {
-        // Required empty constructor
-    }
-
-    public Actor(String name) {
-        this.name = name;
-    }
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public List<Series> getSeriesList() {
-        return seriesList;
-    }
-
-    public void setSeriesList(List<Series> seriesList) {
-        this.seriesList = seriesList;
-    }
-
-    public void addSeries(Series series) {
-        this.seriesList.add(series);
-    }
-
-    @Override
-    public String toString() {
-        return "Actor{" + "name='" + name + '\'' + '}';
-    }
-}
