--- src/main/java/services/SeriesService.java
@@ -1,6 +1,0 @@
-package services;
-
-public interface SeriesService {
-
-    String getSeries();
-}
